Move `u256` to its own module in the crate
Move `u256` to its own module in the crate and add changes to use it
in the code.
